Consolidate all Resque-related initializers into "4_resque"
