--- config/initializers/resque_mailer.rb
@@ -1,1 +1,0 @@
-Resque::Mailer.excluded_environments = []
